# api/openapi.yaml
openapi: 3.0.3
info:
  title: Loom API
  description: |
    The Loom API provides programmatic access to the agent orchestration system.
    
    ## Authentication
    
    API requests must include an API key in the `X-API-Key` header when authentication is enabled.
    
    ## Security
    
    - All production deployments should use HTTPS
    - PKI/TLS certificates can be configured for mutual TLS authentication
    - API keys should be rotated regularly
    
  version: 1.0.0
  contact:
    name: Loom Support
    url: https://github.com/jordanhubbard/loom
servers:
  - url: https://localhost:8443
    description: HTTPS server (production)
  - url: http://localhost:8080
    description: HTTP server (development only)
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/beads:
    get:
      summary: List beads
      description: Get a list of beads, optionally filtered by status, priority, or project
      parameters:
        - name: project_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [open, in_progress, blocked, closed]
        - name: type
          in: query
          schema:
            type: string
            enum: [task, decision, epic]
      responses:
        '200':
          description: List of beads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bead'

    post:
      summary: Create a new bead
      description: Create a new work item or decision bead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BeadCreate'
      responses:
        '201':
          description: Bead created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bead'
  /api/v1/beads/{id}:
    get:
      summary: Get bead details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Bead details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bead'

    patch:
      summary: Update bead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BeadUpdate'
      responses:
        '200':
          description: Bead updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bead'
...
  /api/v1/decisions:
    get:
      summary: List decision beads
      description: Get all decision beads that need resolution
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, claimed, resolved]
        - name: priority
          in: query
          schema:
            type: integer
            enum: [0, 1, 2, 3]
      responses:
        '200':
          description: List of decision beads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DecisionBead'
...
  /api/v1/file-locks:
    get:
      summary: List file locks
      description: Get all current file locks
      parameters:
        - name: project_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of file locks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileLock'

    post:
      summary: Request file lock
      description: Request exclusive access to a file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - file_path
                - project_id
                - agent_id
              properties:
                file_path:
                  type: string
                project_id:
                  type: string
                agent_id:
                  type: string
                bead_id:
                  type: string
      responses:
        '201':
          description: Lock acquired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileLock'
        '409':
          description: File already locked
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Bead:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [task, decision, epic]
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in_progress, blocked, closed]
        priority:
          type: integer
          enum: [0, 1, 2, 3]
        project_id:
          type: string
        assigned_to:
          type: string
        blocked_by:
          type: array
          items:
            type: string
        blocks:
          type: array
          items:
            type: string
        related_to:
          type: array
          items:
            type: string
        parent:
          type: string
        children:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        context:
          type: object
          additionalProperties:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        closed_at:
          type: string
          format: date-time
    BeadCreate:
      type: object
      required:
        - title
        - project_id
      properties:
        type:
          type: string
          enum: [task, decision, epic]
          default: task
        title:
          type: string
        description:
          type: string
        priority:
          type: integer
          enum: [0, 1, 2, 3]
          default: 2
        project_id:
          type: string
        parent:
          type: string
        tags:
          type: array
          items:
            type: string
        context:
          type: object
          additionalProperties:
            type: string
    BeadUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [open, in_progress, blocked, closed]
        assigned_to:
          type: string
        description:
          type: string
        context:
          type: object
          additionalProperties:
            type: string
    DecisionBead:
      allOf:
        - $ref: '#/components/schemas/Bead'
        - type: object
          properties:
            question:
              type: string
            options:
              type: array
              items:
                type: string
            recommendation:
              type: string
            requester_id:
              type: string
            decider_id:
              type: string
            decision:
              type: string
            rationale:
              type: string
            decided_at:
              type: string
              format: date-time
    FileLock:
      type: object
      properties:
        file_path:
          type: string
        project_id:
          type: string
        agent_id:
          type: string
        bead_id:
          type: string
        locked_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
...